Add files via upload
Credit card fraud is a pervasive and financially burdensome issue, resulting in the loss of billions of euros annually. The utilisation of machine learning methodologies has promise in enhancing the efficacy of credit card fraud detection. Nevertheless, the problem of class imbalance in credit card transaction data is a formidable obstacle for machine learning models.
The present study investigates the efficacy of various machine learning models in the identification of fraudulent credit card transactions, while taking into account the challenge posed by class imbalance. This study utilises a dataset comprising credit card transaction data in order to construct and assess different machine learning models. The models were initially constructed without incorporating Principal Component Analysis (PCA) in order to evaluate the influence of class imbalance on the accuracy of the models. Following this, Principal Component Analysis (PCA) is employed to convert the numerical attributes within the dataset into principal components. Consequently, novel machine learning models are constructed utilising the features that have been changed by PCA.
The research demonstrates that the utilisation of machine learning models can yield positive outcomes in the identification of fraudulent credit card transactions, even in scenarios where class imbalance is not explicitly accounted for. Nevertheless, the efficacy of machine learning models can be enhanced by the utilisation of Principal Component Analysis (PCA) on the numerical attributes inside the dataset. The research additionally demonstrates that the Random Forest model attains the most elevated AUC scores in the domain of fraud detection when employing PCA-transformed variables.
Furthermore, the study offers additional insights on the following aspects:
This study examines the influence of various machine learning methods on the accuracy and precision of models in situations when class imbalance is not adequately accounted for. This paper examines the challenges and constraints that arise when employing machine learning models to address the issue of imbalanced datasets in credit card fraud detection. This study investigates the impact of Principal Component Analysis (PCA) transformation on the efficacy of machine learning models in addressing the issue of imbalanced datasets for credit card fraud detection.
This study enhances the current body of research on credit card fraud detection by offering comprehensive insights into the efficacy of various machine learning models, the influence of class imbalance on model performance, and the utility of PCA transformation in enhancing the performance of machine learning models on datasets with imbalanced class distribution.